requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=6.1.1
openpyxl>=3.1.0
numpy>=1.24.0
reportlab>=4.0.0
kaleido>=1.1.0

# IFC processing
ifcopenshell>=0.7.0
shapely>=2.0.0

# Database and persistence
supabase>=2.0.0
python-dotenv>=1.0.0
